Demodulize handlers and priority_map
