+++ lib/tom_queue/delayed_job/external_messages.rb
@@ -19,8 +19,8 @@ module TomQueue
         def resolve_external_handler(work)
 
           # Look for a matching source exchange!
-          klass = TomQueue::DelayedJob.handlers.find { |klass| klass.claim_work?(work) }
-          
+          klass = TomQueue.handlers.find { |klass| klass.claim_work?(work) }
+
           if klass
             debug { "Resolved external handler #{klass} for message. Calling the init block." }
 
@@ -45,7 +45,7 @@ module TomQueue
         # queue_manager - TomQueue::QueueManager to configure against
         #
         def setup_external_handler(queue_manager)
-          TomQueue::DelayedJob.handlers.each do |klass|
+          TomQueue.handlers.each do |klass|
             klass.setup_binding(queue_manager)
           end
         end
