+++ spec/tom_queue/delayed_job/delayed_job_integration_spec.rb
@@ -88,8 +88,8 @@ describe Delayed::Job, "integration spec", :timeout => 10 do
   end
 
   it "should support job priorities" do
-    TomQueue::DelayedJob.priority_map[0] = TomQueue::NORMAL_PRIORITY
-    TomQueue::DelayedJob.priority_map[1] = TomQueue::HIGH_PRIORITY
+    TomQueue.priority_map[0] = TomQueue::NORMAL_PRIORITY
+    TomQueue.priority_map[1] = TomQueue::HIGH_PRIORITY
     Delayed::Job.enqueue(TestJobClass.new("low1"), :priority => 0)
     Delayed::Job.enqueue(TestJobClass.new("low2"), :priority => 0)
     Delayed::Job.enqueue(TestJobClass.new("high"), :priority => 1)
